+++ NEWS.md
@@ -1,5 +1,7 @@
 # gtsummary (development version)
 
+* `tbl_uvregression()` now accepts both data frames and survey design objects as input. (#742)
+
 * Added function `add_vif()` to include variance inflation factors in `tbl_regression()` output. (#717)
 
 # gtsummary 1.3.6